third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/font-style-expected.txt

This is a testharness.js-based test.
[FAIL] Can set 'font-style' to the 'oblique' keyword: oblique
  assert_equals: expected "oblique" but got "italic"
Harness: the test ran to completion.
